change controler a bit

+++ src/GreyScaleAnalysisControler.cpp
@@ -20,12 +20,12 @@ GreyScaleAnalysisControler::~GreyScaleAnalysisControler() {
 void GreyScaleAnalysisControler::print2Console() {
 	cout<< "print2Console method called" << endl;
 	cout<< "print origins" <<endl;
-	cout<< "x: " << this->origins.at<double>(0, 0) <<" y: " << this->origins.at<double>(0, 1) << endl;
-	cout<< "x: " << this->origins.at<double>(1, 0) <<" y: " << this->origins.at<double>(1, 1) << endl;
+	cout<< "x: " << this->origins.at<float>(0, 0) <<" y: " << this->origins.at<float>(0, 1) << endl;
+	cout<< "x: " << this->origins.at<float>(1, 0) <<" y: " << this->origins.at<float>(1, 1) << endl;
 	//cout<< "x: " << this->origins.at<double>(2, 0) <<" y: " << this->origins.at<double>(2, 1) << endl;
 	cout<< "print dimensions" <<endl;
-	cout<< "x: " << this->dimensions.at<double>(0, 0) <<" y: " << this->dimensions.at<double>(0, 1) << endl;
-	cout<< "x: " << this->dimensions.at<double>(1, 0) <<" y: " << this->dimensions.at<double>(1, 1) << endl;
+	cout<< "x: " << this->dimensions.at<float>(0, 0) <<" y: " << this->dimensions.at<float>(0, 1) << endl;
+	cout<< "x: " << this->dimensions.at<float>(1, 0) <<" y: " << this->dimensions.at<float>(1, 1) << endl;
 	//cout<< "x: " << this->dimensions.at<double>(2, 0) <<" y: " << this->dimensions.at<double>(2, 1) << endl;
 }
 
@@ -40,8 +40,8 @@ Mat GreyScaleAnalysisControler::get_histogram() {
 		for (int j = 0; j < n_origins; j++) {
 			string filename = this->get_filename(i, this->file_format);
 			cout << "Processing " + filename + " ROI " + to_string(j) << endl;
-			double origin[2] = {(double)this->origins.at<float>(j,0), (double)this->origins.at<float>(j,1)};
-			double dimension[2] = {(double)this->dimensions.at<float>(j,0), (double)this->dimensions.at<float>(j,1)};
+			double origin[2] = {this->origins.at<float>(j,0), this->origins.at<float>(j,1)};
+			double dimension[2] = {this->dimensions.at<float>(j,0), this->dimensions.at<float>(j,1)};
 			GreyScaleCalculator calculator("../ImageData/" + filename, origin, dimension);
 			calculator.printInfo();
 			double calc_result = calculator.CUDA_greyscale();
